requirements.txt
fastapi[standard]>=0.115.13,<0.116.0
pydantic>=2.11.7,<3.0.0
tinydb>=4.8.2,<5.0.0
python-dotenv>=1.1.0,<2.0.0
requests>=2.32.4,<3.0.0
langchain>=0.3.25,<0.4.0
openai>=1.88.0,<2.0.0
langchain-community>=0.3.25,<0.4.0
langchain-openai>=0.3.24,<0.4.0
auth0-fastapi-api>=1.0.0b3,<2.0.0
auth0-ai-langchain>=0.2.0,<0.3.0 